+++ game/content/ghplots/campfeatures.py
@@ -76,9 +76,9 @@ class MetrosceneRecoveryHandler(Plot):
     def METROSCENE_ENTER(self, camp: gears.GearHeadCampaign):
         # Upon entering this scene, deal with any dead or incapacitated party members.
         # Also, deal with party members who have lost their mecha. This may include the PC.
+        etlr = plotutility.EnterTownLanceRecovery(camp, self.elements["METROSCENE"], self.elements["METRO"])
         if not camp.is_unfavorable_to_pc(self.elements["METROSCENE"]):
             camp.home_base = self.elements["MISSION_GATE"]
-            etlr = plotutility.EnterTownLanceRecovery(camp, self.elements["METROSCENE"], self.elements["METRO"])
 
             if camp.campdata.get(LANCEDEV_ENABLED, False) and random.randint(1, 3) == 2 and not etlr.did_recovery:
                 # We can maybe load a lancemate scene here. Yay!