+++ gears/pets.py
@@ -1,4 +1,5 @@
 class PetData():
     def __init__(self, trainer):
         self.trainer = trainer
+        self.active = True